// src/app/definitions.d.ts
export type Coordinate = [number, number]; // [y, x]
export type GridHash = string; // Représentation de la grille sous forme de chaîne de caractères (ex: ["100", "110", "000"] => "100110000")
// Les web workers ne peuvent recevoir en message que des objets composés de certains types (primitifs, Map, Set, Array, etc.)
export interface RawSimpleGameInstance {
  grid: Array<Array<number>>;
  turn: number;
  moves: Array<Coordinate>;
}

export interface RawScoredGameInstance extends RawSimpleGameInstance {
  score: RawScore;
  nextInstances: Map<number, GridHash>;
}

export interface RawScore {
  score: number; 
  isWinCountdown: boolean; // true si le score est un nombre de coups avant la victoire
  // Lorsque le Score indique un joueur qui a gagné, isWinCountdown = true et score = 0.5 si le joueur 1 a gagné, -0.5 si le joueur 2 a gagné 
}